Split Parse.hs into three seperate modules
