+++ src/Lpt/Parse/Encode.hs
@@ -1,0 +1,1 @@
+module Parse.Encode where
